# app/Http/Controllers/TiposMedicamentosController.php
<?php

namespace App\Http\Controllers;

use App\Models\TiposMedicamentos;
use App\Http\Requests\TipoMedicamentoRequest;
use Illuminate\Http\Request;

/**
 * Class TiposMedicamentosController
 *
 * @autor Vinícius Sarmento Costa Siqueira
 * @link https://github.com/ViniciusSCS
 * @date 13/07/2023 - 13:51
 * @package App\Http\Controllers
 */
class TiposMedicamentosController extends Controller
{
    /**
     * @OA\Post(
     *     tags={"TipoMedicamento"},
     *     path="/tipoMedicamento/cadastro",
     *     security={{"bearerAuth": {}}},
     *     @OA\Parameter(
     *         name="descricao",
     *         required=true,
     *     ),
     *     @OA\Response(response="200", description="Cadastra os Tipos de medicamentos"),
     *     @OA\Response(response="401", description="Usuário não Autenticado"),
     *     @OA\Response(response="422", description="Erro em algum campo obrigatório"),
     * )
     */
    public function create(TipoMedicamentoRequest $request)
    {
        $data = $request->all();

        $tipoMedicamento = new TiposMedicamentos();

        $tipoMedicamento->descricao = $data['descricao'];

        $tipoMedicamento->save();

        return ['status' => true, "tiposMedicamentos" => $tipoMedicamento];
    }

    /**
     * @OA\Get(
     *      tags={"TipoMedicamento"},
     *      path="/tipoMedicamento/listar/",
     *      security={{"bearerAuth": {}}},
     *      @OA\Response(response="200", description="Lista os tipos de medicamentos cadastrados"),
     *      @OA\Response(response="401", description="Usuário não Autenticado"),
     * )
     */
    public function list()
    {
        $query = TiposMedicamentos::select('descricao')->get();

        return ['status' => true, "tiposMedicamentos" => $query];
    }
}