# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistema PetShop",
        "version": "2.02.3"
    },
    "servers": [
        {
            "url": "http://localhost/api"
        }
    ],
    "paths": {
...
        "/raca/select/{id}": {
            "get": {
                "tags": [
                    "Raca"
                ],
                "operationId": "3210523ba4683dd26a2c8c7178f3f829",
                "parameters": [
                    {
                        "name": "id",
                        "in": "/select/{id}",
                        "description": "Especie id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Apreseta a Raça selecionada de acordo com o ID da espécie escolhida"
                    },
                    "401": {
                        "description": "Usuário não Autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tipoMedicamento/cadastro": {
            "post": {
                "tags": [
                    "TipoMedicamento"
                ],
                "operationId": "fc267eab5f6ba548a7bbb04185b788ea",
                "parameters": [
                    {
                        "name": "descricao",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cadastra os Tipos de medicamentos"
                    },
                    "401": {
                        "description": "Usuário não Autenticado"
                    },
                    "422": {
                        "description": "Erro em algum campo obrigatório"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tipoMedicamento/listar/": {
            "get": {
                "tags": [
                    "TipoMedicamento"
                ],
                "operationId": "12c4946ab4bae33bd9a8718b07d56816",
                "responses": {
                    "200": {
                        "description": "Lista os tipos de medicamentos cadastrados"
                    },
                    "401": {
                        "description": "Usuário não Autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "operationId": "e8c53dec7a99697063f337ac03268dfe",
                "parameters": [
                    {
                        "name": "email",
                        "required": true
                    },
                    {
                        "name": "password",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuário loga com email e senha"
                    },
                    "422": {
                        "description": "Erro em algum campo obrigatório"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "User"
                ],
                "operationId": "a1235dd5154ae476b05ba902dc817657",
                "responses": {
                    "200": {
                        "description": "Rota para deslogar usuário e revogar token"
                    },
                    "401": {
                        "description": "Usuário não Autenticado"
                    },
                    "422": {
                        "description": "Erro em algum campo obrigatório"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/cadastro": {
            "post": {
                "tags": [
                    "User"
                ],
                "operationId": "5626d3474d87206d10a628ccb209c7c3",
                "parameters": [
                    {
                        "name": "name",
                        "required": true
                    },
                    {
                        "name": "email",
                        "required": true
                    },
                    {
                        "name": "tipo_id",
                        "required": true
                    },
                    {
                        "name": "password",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cadastra as informações do usuário"
                    },
                    "401": {
                        "description": "Usuário não Autenticado"
                    },
                    "422": {
                        "description": "Erro em algum campo obrigatório"
                    }
                }
            }
        },
        "/user/": {
            "get": {
                "tags": [
                    "User"
                ],
                "operationId": "62dde2fe594c7e364539fa02e378c8be",
                "responses": {
                    "200": {
                        "description": "Apreseta informações do usuário logado"
                    },
                    "401": {
                        "description": "Usuário não Autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/edit": {
            "put": {
                "tags": [
                    "User"
                ],
                "operationId": "17cc4fc658878ec5b4d9884c2bacf646",
                "parameters": [
                    {
                        "name": "name",
                        "required": true
                    },
                    {
                        "name": "email",
                        "required": true
                    },
                    {
                        "name": "tipo_id",
                        "required": true
                    },
                    {
                        "name": "password",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Atualia as informações do usuário logado"
                    },
                    "401": {
                        "description": "Usuário não Autenticado"
                    },
                    "422": {
                        "description": "Erro em algum campo obrigatório"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/deletar": {
            "get": {
                "tags": [
                    "User"
                ],
                "operationId": "be6a9e3a324ead4f6f0e0f4b9aef8d49",
                "responses": {
                    "200": {
                        "description": "Apreseta informações do usuário logado"
                    },
                    "401": {
                        "description": "Usuário não Autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/vacina/select/": {
            "get": {
                "tags": [
                    "Vacina"
                ],
                "operationId": "bc584fd419c87d3cb04aea82e5b30db8",
                "responses": {
                    "200": {
                        "description": "Apreseta todas as Vacinas cadastradas"
                    },
                    "401": {
                        "description": "Usuário não Autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}